[DEV-1620] add ikna controller etc
