+++ src/integration/ikna/dto/ikna-sanction-result.dto.ts
@@ -1,0 +1,8 @@
+import { IknaAddressTag } from './ikna-address-tag.dto';
+
+export interface IknaSanctionResult {
+  testedAddress: string;
+  isSanctioned: boolean | null;
+  sanctionedAddress?: string;
+  sanctionedAddressTags?: IknaAddressTag[];
+}
