+++ src/shared/shared.module.ts
@@ -36,7 +36,6 @@ import { SettingService } from './models/setting/setting.service';
 import { RepositoryFactory } from './repositories/repository.factory';
 import { ApiKeyService } from './services/api-key.service';
 import { HttpService } from './services/http.service';
-import { IknaService } from './services/ikna.service';
 import { PaymentInfoService } from './services/payment-info.service';
 
 @Module({
@@ -70,7 +69,6 @@ import { PaymentInfoService } from './services/payment-info.service';
     ApiKeyService,
     PaymentInfoService,
     IpLogService,
-    IknaService,
   ],
   exports: [
     RepositoryFactory,
@@ -87,7 +85,6 @@ import { PaymentInfoService } from './services/payment-info.service';
     ApiKeyService,
     PaymentInfoService,
     IpLogService,
-    IknaService,
   ],
 })
 export class SharedModule {}
